feat: add convert24HourTimePrefix option
This commit adds a feature that alters the display/visual text of a task
when it is prefixed with a timestamp in 24-hour format. It must be 5
characters in the format "09:00" or "23:45". When this setting is
enabled, any 24-hour timestamp at the very beginning of a task will be
visually converted into 12-hour format with an "am" or "pm" after. This
allows tasks to be sortable by time (24-hour format is sortable, 12-hour
format is not) while having the option to display the times in the
12-hour format if the user finds it easier to read.